refactor: Goパッケージ構成を責務ごとに分離
- internal/mslist/ を新設し、model/types.goからMSリスト関連関数を移動
- internal/pipeline/ を新設し、server.goからJob型とrunPipelineを移動
- model/types.go を型定義のみに整理
- scraper.goから未使用の集計関数(GetDailyScores等)を削除
- FillMsNames/CheckUnknownMS をメソッドからmslistパッケージの関数に変換
- CLAUDE.mdとREADME.mdのコード構成を更新
